add query encoder ance
