+++ pyserini/dsearch/__main__.py
@@ -20,7 +20,7 @@ import os
 import json
 from tqdm import tqdm
 
-from pyserini.dsearch import SimpleDenseSearcher, TCTColBERTQueryEncoder, QueryEncoder, DPRQueryEncoder
+from pyserini.dsearch import SimpleDenseSearcher, TCTColBERTQueryEncoder, QueryEncoder, DPRQueryEncoder, AnceQueryEncoder
 from pyserini.query_iterator import QUERY_IDS, query_iterator
 from pyserini.search import get_topics
 from pyserini.search.__main__ import write_result, write_result_max_passage
@@ -58,6 +58,8 @@ def init_query_encoder(encoder, topics_name, device):
             return DPRQueryEncoder(encoder_dir=encoder, device=device)
         elif 'tct_colbert' in encoder:
             return TCTColBERTQueryEncoder(encoder_dir=encoder, device=device)
+        elif 'ance' in encoder:
+            return AnceQueryEncoder(encoder_dir=encoder, device=device)
     if topics_name in encoded_queries:
         return QueryEncoder.load_encoded_queries(encoded_queries[topics_name])
     return None
